[rhi] Vulkan implementation of scissor rects

+++ include/nova_renderer/rhi/command_list.hpp
@@ -138,7 +138,7 @@ namespace nova::renderer::rhi {
          */
         virtual void draw_indexed_mesh(uint32_t num_indices, uint32_t offset = 0, uint32_t num_instances = 1) = 0;
 
-        virtual void set_scissor_rect(float x, float y, float width, float height) = 0;
+        virtual void set_scissor_rect(uint32_t x, uint32_t y, uint32_t width, uint32_t height) = 0;
 
         virtual ~CommandList() = default;
     };
